TASK-6: Separando acceso a datos de views
